+++ lib/cogs/lastfm.py
@@ -120,7 +120,8 @@ class LastFM(Cog):
     
     @lastfm.group(name="top")
     async def top_group(self, ctx):
-        if ctx.invoked_subcommand != self.top_albums_command and ctx.invoked_subcommand != self.top_tracks_command and ctx.invoked_subcommand != self.top_artist_command:
+        subcommand = ctx.invoked_subcommand
+        if subcommand != self.top_albums_command and subcommand != self.top_tracks_command and subcommand != self.top_artist_command:
             prefix = db.record("SELECT Prefix from guilds WHERE GuildID = ?", ctx.guild.id)
             await ctx.send(f"Try these commands instead.\n`{prefix[0]}fm top albums`\n`{prefix[0]}fm top tracks`\n`{prefix[0]}fm top artists`")
 
@@ -240,7 +241,8 @@ class LastFM(Cog):
 
     @lastfm.group(name="artist")
     async def artist_group(self, ctx):
-        if ctx.invoked_subcommand != self.artist_charts_command:
+        subcommand = ctx.invoked_subcommand
+        if subcommand != self.artist_charts_command and subcommand != self.artist_search_command:
             prefix = db.record("SELECT Prefix from guilds WHERE GuildID = ?", ctx.guild.id)
             await ctx.send(f"Try these commands instead.\n`{prefix[0]}fm artist charts`")
 
@@ -261,6 +263,64 @@ class LastFM(Cog):
                             description='\n'.join(llist), colour=ctx.author.colour)
 
                 await ctx.send(embed=embed)
+
+    @artist_group.command(name="search")
+    async def artist_search_command(self, ctx, *, artist: str):
+        info_url = f"https://ws.audioscrobbler.com/2.0/?method=artist.getinfo&artist={artist}&api_key={os.environ.get('lastfmkey')}&format=json"
+        toptracks = f"https://ws.audioscrobbler.com/2.0/?method=artist.gettoptracks&artist={artist}&api_key={os.environ.get('lastfmkey')}&format=json&limit=5"
+        topalbums = f"https://ws.audioscrobbler.com/2.0/?method=artist.gettopalbums&artist={artist}&api_key={os.environ.get('lastfmkey')}&format=json&limit=5"
+        # artist_pfp_url = f""
+        # if anyone knows an API for getting artist's pfp without needing a unique ID
+        # pls let me know on Discord (mmatt#0001)
+        # Last.fm's api sucks, and doesn't grab them correctly.
+
+        async with request("GET", info_url) as response:
+            data = (await response.json())['artist']
+
+            embed=Embed(title=f"{data['name']} info on Last.fm",
+                        description=data['url'], colour=ctx.author.colour)
+
+            fields = [("Listeners:", data['stats']['listeners'], True),
+                      ("Play Count:", data['stats']['playcount'], True),
+                      ("Wiki:", data['bio']['links']['link']['href'], False)]          
+
+            for name, value, inline in fields:
+                embed.add_field(name=name, value=value, inline=inline)
+
+            similar = list()
+
+            for i in data['similar']['artist']:
+                similar.append(i['name'])
+
+            if similar:
+                embed.add_field(name="Similar Artists:", value=', '.join(similar))
+
+            async with request("GET", toptracks) as response:
+                toptracks = (await response.json())['toptracks']
+
+                tracks = list()
+
+                for i in toptracks['track']:
+                    tracks.append(f"‣ #{i['@attr']['rank']} {i['name']}: Play Count ‣ {i['playcount']}")
+
+                if tracks:
+                    embed.add_field(name="Top Tracks:", value='\n'.join(tracks), inline=False)
+
+            async with request("GET", topalbums) as response:
+                topalbums = (await response.json())['topalbums']
+
+                albums = list()
+
+                for i in topalbums['album']:
+                    albums.append(f"‣ {i['name']}: Play Count ‣ {i['playcount']}")
+
+                if albums:
+                    embed.add_field(name="Top Albums:", value='\n'.join(albums), inline=False)
+  
+            embed.set_thumbnail(url=data['image'][3]['#text'])
+
+            await ctx.send(embed=embed)
+            
 
     @command(name="setlastfm", aliases=["setfm"], brief="Sets your Last.fm username.")
     @cooldown(1, 5, BucketType.user)